fix(gdn): align pre-GDR beta precision
(cherry picked from commit b3dc211171878da57d7ced6e541378a3ed9d0a01)

+++ megatron/core/ssm/gated_delta_net/gdn.py
@@ -121,7 +121,7 @@ class GatedDeltaNet(SSMDynamicInferenceMixin, _GDNBase):
         # ``gate_feats`` arrives in ``in_proj_split_names`` order: beta, then alpha.
         beta, alpha = gate_feats
         g = -A_log_local_cp.exp() * F.softplus(alpha.float() + dt_bias_local_cp)  # In fp32
-        beta = beta.sigmoid()
+        beta = beta.float().sigmoid()
         return g, {"beta": beta.contiguous()}
 
     @jit_fuser
